move settings objs to subdir
